+++ runtime/datatypes/email.ts
@@ -1,5 +1,6 @@
 import * as Red from "../../red-types";
 import RedActions from "../actions";
+import {insert} from "./string-ref";
 
 /* Actions */
 export function $$form(
@@ -8,7 +9,7 @@ export function $$form(
 	buffer: string[],
 	_part?: number
 ): boolean {
-	buffer.push((email.user + "@" + email.host).slice(email.index - 1));
+	buffer.push(email.email.ref.slice(email.index - 1));
 	return false;
 }
 
@@ -20,4 +21,17 @@ export function $$mold(
 	_: RedActions.MoldOptions = {}
 ): boolean {
 	return $$form(ctx, email, buffer, _.part);
+}
+
+// ...
+
+export function $$insert(
+	ctx:   Red.Context,
+	email: Red.RawEmail,
+	value: Red.AnyType,
+	_: RedActions.InsertOptions = {}
+): Red.RawEmail {
+	email.index += insert(ctx, email.email, email.index - 1, value, _, encodeURI);
+	
+	return email;
 }